Moved a lot of stream related stuff out of the public headers and replaced more ostream dependencies with iosfwd
