--- include/internal/catch_streambuf.cpp
@@ -1,12 +1,0 @@
-/*
- *  Created by Martin on 31/08/2017.
- *
- *  Distributed under the Boost Software License, Version 1.0. (See accompanying
- *  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
- */
-
-#include "catch_streambuf.h"
-
-namespace Catch {
-    StreamBufBase::~StreamBufBase() = default;
-}
